+++ pkgs/dart_mcp_server/lib/src/arg_parser.dart
@@ -78,7 +78,6 @@ ArgParser createArgParser({
               'disabled category to re-enable more specific features under '
               'that category.',
           allowed: allFeatureAndCategoryNames,
-          defaultsTo: [FeatureCategory.all.name],
         )
         ..addOption(
           toolsOption,