+++ pkgs/dart_mcp_server/test/test_harness.dart
@@ -98,6 +98,7 @@ class TestHarness {
     bool forceRootsFallback = false,
     Sdk? sdk,
     bool startFakeEditorExtension = true,
+    FeaturesConfiguration featuresConfig = const FeaturesConfiguration(),
   }) async {
     sdk ??= Sdk.find(
       dartSdkPath: Platform.environment['DART_SDK'],
@@ -117,6 +118,7 @@ class TestHarness {
       sdk,
       cliArgs,
       forceRootsFallback,
+      featuresConfig,
     );
     final connection = serverConnectionPair.serverConnection;
     connection.onLog.listen((log) {
@@ -475,6 +477,7 @@ Future<ServerConnectionPair> _initializeMCPServer(
   Sdk sdk,
   List<String> cliArgs,
   bool forceRootsFallback,
+  FeaturesConfiguration featuresConfig,
 ) async {
   ServerConnection connection;
   DartMCPServer? server;
@@ -519,13 +522,11 @@ Future<ServerConnectionPair> _initializeMCPServer(
 
     server = DartMCPServer(
       serverChannel,
-      featuresConfig: FeaturesConfiguration(),
+      featuresConfig: featuresConfig,
       processManager: processManager,
       fileSystem: fileSystem,
       sdk: sdk,
       analytics: analytics,
-      // So we can test them.
-      enableScreenshots: true,
       forceRootsFallback: forceRootsFallback,
     );
     addTearDown(server.shutdown);
@@ -541,6 +542,8 @@ Future<ServerConnectionPair> _initializeMCPServer(
       'run',
       'bin/main.dart',
       ...cliArgs,
+      for (var enabled in featuresConfig.enabledNames) '--enable=$enabled',
+      for (var disabled in featuresConfig.disabledNames) '--disable=$disabled',
     ]);
     addTearDown(process.kill);
     connection = client.connectServer(